Correct and refactor integration test support. Added an start and stop goal, additional to the already existing run goal.
Enhance environment variable support, i.e.
Maven properties are now supported.
